// TravellerCreator/ClientApp/src/app/models/sector-creator/hex/system/planet/planet-type.ts
export enum PlanetType {
  Acheronian = "Acheronian",
  Arean = "Arean",
  Arid = "Arid",
  Asphodelian = "Asphodelian",
  AsteroidBelt = "AsteroidBelt",
  Chthonian = "Chthonian",
  Dwarf = "Dwarf",
  Hebean = "Hebean",
  Helian = "Helian",
  JaniLithic = "JaniLithic",
  Jovian = "Jovian",
  Meltball = "Meltball",
  Oceanic = "Oceanic",
  Panthalassic = "Panthalassic",
  Promethean = "Promethean",
  Rockball = "Rockball",
  SmallBody = "SmallBody",
  Snowball = "Snowball",
  Stygian = "Stygian",
  Tectonic = "Tectonic",
  Telluric = "Telluric",
  Terrestrial = "Terrestrial",
  Vesperian = "Vesperian",
  Unknown = "Unknown",
  Scrap = "Scrap",
}