// TravellerCreator/ClientApp/src/app/models/sector-creator/hex/system/planet/settlement/base.ts
export enum Base {
  navalDepot = "Nd",
  prison = "Pr",
  naturePreserve = "Pr",
  foreignEmbassy = "Em",
  imperialConsulate = "Co",
  governorsEstate = "Go",
  hospital = "Ho",
  galacticChurch = "Ch",
  dataRepository = "Da",
  merchantBase = "Me",
  megacorporateBase = "Mc",
  navalBase = "Na",
  pirateBase = "Pi",
  ancientSite = "An",
  researchStation = "Re",
  scoutBase = "Sc",
  scoutHostel = "Sh",
  TASHostel = "Tc",
  TASFirstClassHostel = "Tb",
  TASChapterHouse = "Ta",
  galacticUniversity = "Un",
  explorationBase = "Ex",
  terraformingFacility = "Tf",
  shipyard = "Sh",
  psionicsInstitute = "Ps",
  wayStation = "Wa",
}